renamed some routes, removed other
