+++ routes/index.js
@@ -2,9 +2,7 @@ const router = require('express').Router()
 
 router.use('/api', require('./userRoutes.js'))
 
-router.use('/api', require('./actualRoutes.js'))
-
-router.use('/api', require('./goalRoutes.js'))
+router.use('/api', require('./monthRoutes.js'))
 
 router.use('/api', require('./categoryRoutes.js'))
